Data is Pending to load in Childs

+++ demo/models.py
@@ -5,7 +5,7 @@ from django.db import models
 
 class Parents(models.Model):
     name = models.CharField(max_length=100)
-    mobile = models.IntegerField()
+    mobile = models.BigIntegerField()
     gender = models.CharField(max_length=30)
     age = models.IntegerField()
